Card Picker Auto Layout (#29)
* Layout

It lays the cards out depending on how many cards there are.

Not perfect. Only works perfectly on some devices. This needs fixing.

* Weird Layout

Its weird on some phones. I tested them all and it shows in a commented out line which ones are messed up and how they are messed up.

* Jokers

Added Jokers to the cards array.

* Auto Layout Card Picker

Have it show the card size bigger or smaller depending on the amount of cards they want shuffled.

Tested on all devices, works on all except ipod touch 7th gen(5 and 6 cards break the view)

Co-authored-by: Matthew <matt@Matthews-MacBook-Pro.local>