Removes From Array
when you left the settings page with the thing on it would turn it off automatically and keep the jokers in and if you turned it on again their would be 4 jokers.

Fixed that ^ problem where if you leave it will remove the jokers. Need to make it so that it keeps if on if they left it on.