Dice roller (#57)
* Added Dice Options

They dont work yet but the buttons are there for more dice

* Added View For Dice

Made it layout for up to 9 dice.

They all show the same image(still need to fix this)

* Forgot to Add

Forgot To Add to last commit

* Dice Roller

Made the code cleaner and next need to add buttons to add that specific dice

* forgot to add this to the last commit

* Added Buttons

Added buttons above the roll dice to add a dice. It adds a dice but not the right one yet

* Had To Commit

* Added Dice

Redid the code and added it so it shows the dice that you picked. Just need the pictures now

* Lots of dice

embeded the diceView in a scroll view so that they can roll lots of dice and it wont ruin the layout

* Remove Dice

When you tap on a dice it removes that dice, also when you tap anywhere it removes the first dice. There is a space after 6 dice that needs to be fixed

* cleaned up

just cleaned up some extra space

* testing

had to make this commit to merge

Co-authored-by: Matthew <matt@Matthews-MacBook-Pro.local>